// frontend/src/api/generated/jobs/jobs.ts
export type listUtilityJobsApiV1JobsGetResponse200 = {
  data: LimitOffsetPageTypeVarCustomizedUtilityJobRead;
  status: 200;
};

export type listUtilityJobsApiV1JobsGetResponse422 = {
  data: HTTPValidationError;
  status: 422;
};

export type listUtilityJobsApiV1JobsGetResponseSuccess =
  listUtilityJobsApiV1JobsGetResponse200 & {
    headers: Headers;
  };
export type listUtilityJobsApiV1JobsGetResponseError =
  listUtilityJobsApiV1JobsGetResponse422 & {
    headers: Headers;
  };

export type listUtilityJobsApiV1JobsGetResponse =
  | listUtilityJobsApiV1JobsGetResponseSuccess
  | listUtilityJobsApiV1JobsGetResponseError;
// ...
export type listUtilityJobScriptOptionsApiV1JobsScriptOptionsGetResponse200 = {
  data: UtilityJobScriptOption[];
  status: 200;
};

export type listUtilityJobScriptOptionsApiV1JobsScriptOptionsGetResponseSuccess =
  listUtilityJobScriptOptionsApiV1JobsScriptOptionsGetResponse200 & {
    headers: Headers;
  };
export type listUtilityJobScriptOptionsApiV1JobsScriptOptionsGetResponse =
  listUtilityJobScriptOptionsApiV1JobsScriptOptionsGetResponseSuccess;